we can diverge MP/non-MP in the future
