+++ lmcache/observability.py
@@ -23,7 +23,7 @@ import prometheus_client
 
 # First Party
 from lmcache.logging import init_logger
-from lmcache.usage_telemetry.non_mp import ContinuousUsageContext
+from lmcache.usage_telemetry import ContinuousUsageContext
 from lmcache.utils import thread_safe
 from lmcache.v1.metadata import LMCacheMetadata
