--- lmcache/usage_telemetry/mp/__init__.py
@@ -1,27 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-"""Usage reporting for the multiprocess (MP) cache server.
-
-All MP payloads POST under the ``/mp/`` endpoint prefix (applied by
-``usage_server_url``). Single-process reporting lives in
-:mod:`lmcache.usage_telemetry.non_mp` and is scheduled for removal.
-
-Importing this package pulls in :mod:`lmcache.v1.mp_observability`; only
-MP-server code should import it.
-"""
-
-# First Party
-from lmcache.usage_telemetry.mp.context import (
-    InitializeMPUsageContext,
-    MPUsageContext,
-)
-from lmcache.usage_telemetry.mp.continuous import (
-    InitializeMPContinuousUsage,
-    MPContinuousUsageReporter,
-)
-
-__all__ = [
-    "InitializeMPContinuousUsage",
-    "InitializeMPUsageContext",
-    "MPContinuousUsageReporter",
-    "MPUsageContext",
-]
