--- lmcache/usage_telemetry/non_mp/__init__.py
@@ -1,20 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-"""Usage reporting for the single-process (non-MP) LMCacheEngine path.
-
-Scheduled for removal together with the single-process code path; do not
-add new dependencies on this package. MP-mode reporting lives in
-:mod:`lmcache.usage_telemetry.mp`.
-"""
-
-# First Party
-from lmcache.usage_telemetry.non_mp.context import (
-    InitializeUsageContext,
-    UsageContext,
-)
-from lmcache.usage_telemetry.non_mp.continuous import ContinuousUsageContext
-
-__all__ = [
-    "ContinuousUsageContext",
-    "InitializeUsageContext",
-    "UsageContext",
-]
