+++ lmcache/v1/multiprocess/server.py
@@ -13,10 +13,8 @@ import zmq
 # First Party
 from lmcache import torch_dev, torch_device_type
 from lmcache.logging import init_logger
-from lmcache.usage_telemetry.mp import (
-    InitializeMPContinuousUsage,
-    InitializeMPUsageContext,
-)
+from lmcache.usage_telemetry.mp import InitializeMPUsageContext
+from lmcache.usage_telemetry.mp_continuous import InitializeMPContinuousUsage
 from lmcache.v1.distributed.config import (
     StorageManagerConfig,
     add_storage_manager_args,
